chore: move `/shoot/state/finalizer` to `/shootstate` directory
This commit moves the
`pkg/controllermanager/controller/shoot/state/finalizer` into
`pkg/controllermanager/controller/shootstate` as the controller
is used for `ShootState` reconciliation and it makes more sense
to be outside of the `/shoot` directory.
