+++ pkg/controllermanager/controller/shoot/add.go
@@ -17,7 +17,6 @@ import (
 	"github.com/gardener/gardener/pkg/controllermanager/controller/shoot/quota"
 	"github.com/gardener/gardener/pkg/controllermanager/controller/shoot/reference"
 	"github.com/gardener/gardener/pkg/controllermanager/controller/shoot/retry"
-	"github.com/gardener/gardener/pkg/controllermanager/controller/shoot/state/finalizer"
 	"github.com/gardener/gardener/pkg/controllermanager/controller/shoot/statuslabel"
 )
 
@@ -67,12 +66,6 @@ func AddToManager(mgr manager.Manager, cfg controllermanagerconfigv1alpha1.Contr
 		Config: *cfg.Controllers.ShootStatusLabel,
 	}).AddToManager(mgr); err != nil {
 		return fmt.Errorf("failed adding statuslabel reconciler: %w", err)
-	}
-
-	if err := (&finalizer.Reconciler{
-		Config: *cfg.Controllers.ShootState,
-	}).AddToManager(mgr); err != nil {
-		return fmt.Errorf("failed adding shootState finalizer reconciler: %w", err)
 	}
 
 	return nil
