+++ pkg/controllermanager/controller/shootstate/shootstate_suite_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package finalizer
+package shootstate
 
 import (
 	"testing"
